Move owner commands to new cog
